Function pallet_node_authorization::pallet::dispatchables::remove_connections

pub fn remove_connections<T: Config>(
    node: OpaquePeerId,
    connections: Vec<OpaquePeerId>,
)

Remove additional connections of a given node.

node: identifier of the node.

connections: additional nodes from which the connections are not allowed anymore.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::remove_connections.