Trait pallet_node_authorization::Trait

pub trait Trait: Trait {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type MaxWellKnownNodes: Get<u32>;
    type MaxPeerIdLength: Get<u32>;
    type AddOrigin: EnsureOrigin<Self::Origin>;
    type RemoveOrigin: EnsureOrigin<Self::Origin>;
    type SwapOrigin: EnsureOrigin<Self::Origin>;
    type ResetOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;
}

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The event type of this module.

type MaxWellKnownNodes: Get<u32>

The maximum number of well known nodes that are allowed to set

type MaxPeerIdLength: Get<u32>

The maximum length in bytes of PeerId

type AddOrigin: EnsureOrigin<Self::Origin>

The origin which can add a well known node.

type RemoveOrigin: EnsureOrigin<Self::Origin>

The origin which can remove a well known node.

type SwapOrigin: EnsureOrigin<Self::Origin>

The origin which can swap the well known nodes.

type ResetOrigin: EnsureOrigin<Self::Origin>

The origin which can reset the well known nodes.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Loading content...