Crate pallet_node_authorization

Node authorization pallet

This pallet manages a configurable set of nodes for a permissioned network. Each node is identified by a PeerId (i.e. Vec<u8>). It provides two ways to authorize a node,

a set of well known nodes across different organizations in which the connections are allowed.

users can claim the ownership for each node, then manage the connections of the node.

A node must have an owner. The owner can additionally change the connections for the node. Only one user is allowed to claim a specific node. To eliminate false claim, the maintainer of the node should claim it before even starting the node. This pallet uses offchain worker to set reserved nodes, if the node is not an authority, make sure to enable offchain worker with the right CLI flag. The node can be lagged with the latest block, in this case you need to disable offchain worker and manually set reserved nodes when starting it.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_node_authorization