Module pallet_node_authorization::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

add_connections

Add additional connections to a given node.

add_well_known_node

Add a node to the set of well known nodes. If the node is already claimed, the owner will be updated and keep the existing additional connection unchanged.

claim_node

A given node can be claimed by anyone. The owner should be the first to know its PeerId, so claim it right away!

remove_claim

A claim can be removed by its owner and get back the reservation. The additional connections are also removed. You can’t remove a claim on well known nodes, as it needs to reach consensus among the network participants.

remove_connections

Remove additional connections of a given node.

remove_well_known_node

Remove a node from the set of well known nodes. The ownership and additional connections of the node will also be removed.

reset_well_known_nodes

Reset all the well known nodes. This will not remove the ownership and additional connections for the removed nodes. The node owner can perform further cleaning if they decide to leave the network.

swap_well_known_node

Swap a well known node to another. Both the ownership and additional connections stay untouched.

transfer_node

A node can be transferred to a new owner.